Change Wallpaper to My Little Pony
A prank piece of code, meant to be turned into an executable
so that when plugged into a pc it changes their wallpaper into a My Little Pony,
Henceforth the name of the file CWTM (Change Wallpaper To MLP). 
Only for windows for now. 
A limitation of whether the person has python installed on their machine.
Unfortunately Windows disabled autorun back on windows 7 so the exe file has to be ran manually.